rename serde_utils to serde_snapshot
