+++ runtime/src/lib.rs
@@ -14,7 +14,7 @@ pub mod message_processor;
 mod native_loader;
 pub mod nonce_utils;
 pub mod rent_collector;
-pub mod serde_utils;
+pub mod serde_snapshot;
 pub mod stakes;
 pub mod status_cache;
 mod system_instruction_processor;
